refactor: participans components folder structure
