+++ src/app/events/[event]/participants/page.tsx
@@ -5,8 +5,12 @@ import { Round } from "@/shared/types"
 export default function Participants({ params }: EventProps) {
   return (
     <section className="font-sans">
-      <SubTitle>Code in The Dark {params.event}</SubTitle>
-      <Title>Gerenciar Chave</Title>
+      <p className="text-neutral-500 text-body-xs underline leading-normal">
+        Code in The Dark {params.event}
+      </p>
+      <h2 className="text-title-sm text-neutral-900 leading-normal font-bold mt-1">
+        Gerenciar Chave
+      </h2>
       <RoundsList rounds={rounds} />
     </section>
   )
