+++ src/components/participants/participant-image.tsx
