+++ src/components/participants/participant-item.tsx
