+++ src/components/participants/participants-list.tsx
