--- src/components/participants/participants-list/index.ts
@@ -1,4 +1,0 @@
-export * from "./participants-list"
-export * from "./participant-image"
-export * from "./participant-item"
-export * from "./participant-name"
