--- src/components/participants/participants-list/participant-image/index.ts
@@ -1,1 +1,0 @@
-export * from './participant-image'
