--- src/components/participants/participants-list/participant-item/index.ts
@@ -1,1 +1,0 @@
-export * from './participant-item'
