--- src/components/participants/participants-list/participant-name/index.ts
@@ -1,1 +1,0 @@
-export * from './participant-name'
