--- src/components/participants/sub-title/index.ts
@@ -1,1 +1,0 @@
-export * from "./sub-title"
